requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.4.2
attrs==21.4.0
autopep8==1.5.4
backcall==0.2.0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
colorama==0.4.4
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.3
et-xmlfile==1.1.0
future==0.18.2
idna==3.3
importlib-metadata==4.10.1
importlib-resources==5.4.0
ipykernel==6.7.0
ipython==7.31.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==5.5.1
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
lazy-object-proxy==1.4.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
nbclient==0.5.10
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.7
numpy==1.21.5
openpyxl==3.0.9
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
prometheus-client==0.12.0
prompt-toolkit==3.0.24
pycodestyle==2.6.0
pycparser==2.21
Pygments==2.11.2
pylint==2.6.0
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
PyVimeo==1.1.0
#pywin32
# pywinpty==1.1.6
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.0
requests==2.27.1
Send2Trash==1.8.0
six==1.16.0
terminado==0.12.1
testpath==0.5.0
tinydb==4.6.1
toml==0.10.1
tornado==6.1
traitlets==5.1.1
tuspy==0.2.5
typing_extensions==4.0.1
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
wincertstore==0.2
wrapt==1.12.1
zipp==3.7.0
gunicorn
Werkzeug==2.0.2
Flask
Flask-Cors
# local packages -
-e .
